test(corpus): consolidate extract_schema_names + tighten sagri/mrv test
Code-reviewer + simplifier feedback on commit dd8b2f8:

- extract_schema_names was duplicated verbatim in tests/corpus.rs and
  tests/corpus_sagri_mrv.rs; move to tests/common/mod.rs as the single
  source of truth.
- Drop the redundant standalone CREATE SCHEMA "extensions" call; just
  insert "extensions" into the BTreeSet before collecting and let the
  existing loop create it once.
- Replace .unwrap() on get_host_port_ipv4 with .expect(...) so a failure
  there gives a meaningful message.
- Improve the convergence-failure panic to also print the SQL pgmold
  would generate for the remaining ops, not just the MigrationOp debug
  output. Makes diagnosis actionable when the snapshot regresses.
