+++ tests/corpus_sagri_mrv.rs
@@ -8,13 +8,11 @@
 //! TABLE, GIST indexes).
 //!
 //! Requires PostGIS, so we use the `postgis/postgis` image. Marked `#[ignore]`
-//! so it only runs with `cargo test --test corpus_sagri_mrv -- --ignored` (or
-//! the umbrella corpus test target).
+//! so it only runs with `cargo test --test corpus_sagri_mrv -- --ignored`.
 
 mod common;
 use common::*;
 
-use std::collections::BTreeSet;
 use testcontainers_modules::postgres::Postgres as PostgresImage;
 
 const SNAPSHOT: &str = include_str!("corpus/sagri_mrv.sql");
@@ -26,27 +24,12 @@ async fn setup_postgis() -> (testcontainers::ContainerAsync<PostgresImage>, Stri
         .start()
         .await
         .expect("postgis container should start");
-    let port = container.get_host_port_ipv4(5432).await.unwrap();
+    let port = container
+        .get_host_port_ipv4(5432)
+        .await
+        .expect("get postgis container port");
     let url = format!("postgres://postgres:postgres@localhost:{port}/postgres");
     (container, url)
-}
-
-fn extract_schema_names(sql: &str) -> BTreeSet<String> {
-    let mut schemas = BTreeSet::new();
-    schemas.insert("public".to_string());
-    for line in sql.lines() {
-        let normalized = line.trim().to_uppercase();
-        let rest = normalized
-            .strip_prefix("CREATE SCHEMA IF NOT EXISTS ")
-            .or_else(|| normalized.strip_prefix("CREATE SCHEMA "));
-        if let Some(rest) = rest {
-            let name = rest.trim_end_matches(';').trim().trim_matches('"');
-            if !name.is_empty() && name != "PUBLIC" {
-                schemas.insert(name.to_lowercase());
-            }
-        }
-    }
-    schemas
 }
 
 #[tokio::test]
@@ -55,24 +38,17 @@ async fn sagri_mrv_snapshot_converges() {
     let (_container, url) = setup_postgis().await;
     let connection = PgConnection::new(&url).await.unwrap();
 
-    // Supabase pre-creates the `extensions` schema; the snapshot relies on
-    // that and never issues `CREATE SCHEMA "extensions"`. Create it here so
-    // `CREATE EXTENSION ... WITH SCHEMA extensions` doesn't fail on apply.
-    sqlx::query(r#"CREATE SCHEMA IF NOT EXISTS "extensions""#)
-        .execute(connection.pool())
-        .await
-        .expect("create extensions schema");
-
     sqlx::query("CREATE EXTENSION IF NOT EXISTS postgis")
         .execute(connection.pool())
         .await
         .expect("postgis extension should install");
 
-    let mut schema_names: Vec<String> = extract_schema_names(SNAPSHOT).into_iter().collect();
-    if !schema_names.iter().any(|s| s == "extensions") {
-        schema_names.push("extensions".to_string());
-        schema_names.sort();
-    }
+    // Supabase pre-creates the `extensions` schema; the snapshot relies on
+    // that and never issues `CREATE SCHEMA "extensions"`. Inject it into the
+    // pre-create set so the loop below creates it before apply runs.
+    let mut schemas = extract_schema_names(SNAPSHOT);
+    schemas.insert("extensions".to_string());
+    let schema_names: Vec<String> = schemas.into_iter().collect();
     for schema in &schema_names {
         if schema != "public" {
             sqlx::query(&format!("CREATE SCHEMA IF NOT EXISTS \"{schema}\""))
@@ -102,10 +78,15 @@ async fn sagri_mrv_snapshot_converges() {
         .unwrap();
     let second_diff = compute_diff(&after, &target);
 
-    assert!(
-        second_diff.is_empty(),
-        "sagri_mrv snapshot did not converge: {} op(s) remain after apply: {:#?}",
-        second_diff.len(),
-        second_diff
-    );
+    if !second_diff.is_empty() {
+        let remaining_sql = generate_sql(&plan_migration(second_diff.clone()));
+        panic!(
+            "sagri_mrv snapshot did not converge: {} op(s) remain after apply\n\
+             remaining ops: {:#?}\n\
+             remaining SQL:\n{}",
+            second_diff.len(),
+            second_diff,
+            remaining_sql.join("\n"),
+        );
+    }
 }
